refactored game index
